save metadata for mediawiki downloader
keeping metadata to document the license status of downloaded files,
especially for CC-BY-* and CC0.

metadata is unfortunately almost completely unstructured. it would be
possible to parse the various templates used to generate the "Summary"
section. but these then have stuff like an empty `Permission=` field,
only to then have a `{{Cc-zero}}` tag in the "Licensing" section. so we
instead just save the entire mess without even attempting to parse it.
that's enough to document a file's copyright status, though.